+++ setup.py
@@ -74,6 +74,9 @@ setup(
         'cryptography',
     ],
     extras_require={
+        ':python_version<"3.7"': [
+            'dataclasses',
+        ],
         'kerberos:sys_platform=="win32"': [],
         'kerberos:sys_platform!="win32"': [
             'gssapi>=1.5.0',
