+++ tests/conftest.py
@@ -97,7 +97,7 @@ def ntlm_cred(tmpdir, monkeypatch):
 def kerb_cred(monkeypatch):
     test_kerberos = os.environ.get('PYSPNEGO_TEST_KERBEROS', None)
     if not test_kerberos or not HAS_K5TEST:
-        pytest.skip("Cannot create Kerberos credential without PYSPNEGO_TEST_REALM being set")
+        pytest.skip("Cannot create Kerberos credential without PYSPNEGO_TEST_KERBEROS being set")
 
     realm = K5Realm()
     try:
